Preserve pinned validation during offline installs (#271)
Reject unsafe ref values and unsupported workspace pins before dependency
installation can modify the host.

Keep prebuilt commit validation active when `--no-update` reuses a detached
checkout from an earlier pinned install, without reporting that inherited
state as a newly requested pin. Add regression coverage and split oversized
test modules to retain the repository's module-size limit.
