+++ installer/src/install_flow.rs
@@ -25,6 +25,24 @@ use whitaker_installer::output::write_stderr_line;
 use whitaker_installer::prebuilt::{PrebuiltConfig, PrebuiltResult, attempt_prebuilt};
 use whitaker_installer::prebuilt_path::prebuilt_library_dir;
 use whitaker_installer::resolution::{EXPERIMENTAL_LINT_CRATES, LINT_CRATES, SUITE_CRATE};
+use whitaker_installer::workspace::{WorkspaceAction, ensure_ref_allowed};
+
+/// Validates a ref before invoking the dependency installer.
+pub(crate) fn ensure_dependencies_after_ref_validation<F>(
+    args: &InstallArgs,
+    action: &WorkspaceAction,
+    stderr: &mut dyn Write,
+    ensure_dependencies: F,
+) -> Result<()>
+where
+    F: FnOnce(bool, &mut dyn Write) -> Result<()>,
+{
+    ensure_ref_allowed(action, args.git_ref.as_deref())?;
+    if !args.skip_deps {
+        ensure_dependencies(args.quiet, stderr)?;
+    }
+    Ok(())
+}
 
 pub(crate) fn ensure_dylint_tools_core(
     quiet: bool,
