+++ installer/src/main.rs
@@ -12,7 +12,8 @@ mod workspace_progress;
 use crate::install_flow::ensure_dylint_tools_with_options;
 use crate::install_flow::{
     MetricsWriteContext, PrebuiltInstallationContext, detect_host_target,
-    ensure_dylint_tools_with_executor, try_prebuilt_installation, write_install_metrics,
+    ensure_dependencies_after_ref_validation, ensure_dylint_tools_with_executor,
+    try_prebuilt_installation, write_install_metrics,
 };
 use camino::{Utf8Path, Utf8PathBuf};
 use clap::Parser;
@@ -94,9 +95,8 @@ fn try_fast_path_installation(
 
 /// Runs the install command to build and stage lint libraries.
 ///
-/// Workflow: (1) check/install Dylint dependencies, (2) locate/clone workspace,
-/// (3) resolve crates from CLI flags, (4) build in release mode, (5) stage
-/// libraries with toolchain-suffixed names, (6) generate wrapper script.
+/// Validates the request, prepares dependencies and a workspace, then installs
+/// the selected lint libraries and wrapper scripts.
 ///
 /// # Errors
 ///
@@ -109,14 +109,11 @@ fn run_install(args: &InstallArgs, stderr: &mut dyn Write) -> Result<()> {
         return run_dry(args, &dirs, stderr);
     }
     let install_started = Instant::now();
-    // Step 1: Check and install Dylint dependencies if needed
-    if !args.skip_deps {
-        ensure_dylint_tools(args.quiet, stderr)?;
-    }
-    // Step 2: Ensure workspace is available (clone if needed)
+    // Reject unsafe pin requests before dependency installation can mutate the host.
+    validate_ref_then_ensure_dependencies(args, &dirs, stderr, ensure_dylint_tools)?;
     let workspace = ensure_whitaker_workspace(args, &dirs, stderr)?;
+    let expected_git_sha = workspace.expected_git_sha().map(str::to_owned);
     let workspace_root = workspace.root;
-    // Step 3: Resolve crates and toolchain
     let requested_crates = resolve_requested_crates(args)?;
     let toolchain = resolve_toolchain(&workspace_root, args.toolchain.as_deref())?;
     ensure_toolchain_installed(
@@ -126,14 +123,13 @@ fn run_install(args: &InstallArgs, stderr: &mut dyn Write) -> Result<()> {
         stderr,
     )?;
     let target_dir = determine_target_dir(args.target_dir.as_deref())?;
-    // Step 3.5: Attempt prebuilt download or staged-suite fast path.
     let fast_path_context = FastPathContext {
         args,
         dirs: &dirs,
         requested_crates: &requested_crates,
         toolchain: &toolchain,
         target_dir: &target_dir,
-        expected_git_sha: workspace.pinned_commit.as_deref(),
+        expected_git_sha: expected_git_sha.as_deref(),
     };
     if let Some((staging_path, install_mode)) =
         try_fast_path_installation(&fast_path_context, stderr)?
@@ -156,7 +152,6 @@ fn run_install(args: &InstallArgs, stderr: &mut dyn Write) -> Result<()> {
         experimental: args.experimental,
         quiet: args.quiet,
     };
-    // Step 4: Build and stage
     let build_results = perform_build(&context, &requested_crates, stderr)?;
     let staging_path = stage_libraries(&context, &build_results, stderr)?;
     // Step 5: Generate wrapper scripts if requested
@@ -201,6 +196,22 @@ fn run_dry(args: &InstallArgs, dirs: &dyn BaseDirs, stderr: &mut dyn Write) -> R
     };
     write_stderr_line(stderr, info.display_text());
     Ok(())
+}
+
+/// Validates a pin request before permitting dependency installation.
+fn validate_ref_then_ensure_dependencies<F>(
+    args: &InstallArgs,
+    dirs: &dyn BaseDirs,
+    stderr: &mut dyn Write,
+    ensure_dependencies: F,
+) -> Result<()>
+where
+    F: FnOnce(bool, &mut dyn Write) -> Result<()>,
+{
+    use whitaker_installer::workspace::resolve_workspace_action;
+
+    let action = resolve_workspace_action(dirs, !args.no_update)?;
+    ensure_dependencies_after_ref_validation(args, &action, stderr, ensure_dependencies)
 }
 
 fn determine_dry_run_target_dir(
